+++ lib/textus/manifest/entry/validators.rb
@@ -6,7 +6,6 @@ module Textus
           Events,
           Publish,
           InjectBoot,
-          IndexFilename,
           Ignore,
           FormatMatrix,
         ].freeze